# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SHOP Quote API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quotes:
    get:
      operationId: listQuotes
      tags: [quotes]
      summary: List all quotes
      responses:
        "200":
          description: List of quotes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuoteSummary"
    post:
      operationId: createQuote
      tags: [quotes]
      summary: Create a quote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
  /quotes/{id}:
    get:
      operationId: getQuote
      tags: [quotes]
      summary: Get a quote with line items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
        "404":
          description: Not found
    patch:
      operationId: updateQuote
      tags: [quotes]
      summary: Update a quote
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
    delete:
      operationId: deleteQuote
      tags: [quotes]
      summary: Delete a quote
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /quotes/{id}/duplicate:
    post:
      operationId: duplicateQuote
      tags: [quotes]
      summary: Duplicate a quote
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Duplicated quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
  /quotes/{id}/drawings:
    get:
      operationId: listQuoteDrawings
      tags: [quotes]
      summary: List drawings for a quote
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of drawings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuoteDrawing"
    post:
      operationId: createQuoteDrawing
      tags: [quotes]
      summary: Add a drawing record to a quote
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuoteDrawingBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteDrawing"
  /quotes/{id}/drawings/{drawingId}:
    delete:
      operationId: deleteQuoteDrawing
      tags: [quotes]
      summary: Delete a drawing from a quote
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: drawingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    QuoteLineItemInput:
      type: object
      required:
        - partName
        - quantity
        - material
        - processType
        - setupHours
        - programmingHours
        - machiningMinutesPerPart
        - inspectionHours
        - deburringMinutesPerPart
        - materialCostPerUnit
        - materialWastagePercentage
        - toolingAllowance
        - outsideProcessing
        - packaging
        - delivery
        - riskPercentage
        - profitMarginPercentage
      properties:
        partName:
          type: string
        drawingNumber:
          type: string
        revision:
          type: string
        quantity:
          type: integer
        material:
          type: string
        processType:
          type: string
        machineId:
          type: ["integer", "null"]
        toleranceClass:
          type: string
        surfaceFinish:
          type: string
        complexity:
          type: string
        lineItemType:
          type: string
        hiddenFromPdf:
          type: boolean
        setupHours:
          type: number
        programmingHours:
          type: number
        machiningMinutesPerPart:
          type: number
        inspectionHours:
          type: number
        deburringMinutesPerPart:
          type: number
        materialCostPerUnit:
          type: number
        materialWastagePercentage:
          type: number
        toolingAllowance:
          type: number
        outsideProcessing:
          type: number
        packaging:
          type: number
        delivery:
          type: number
        riskPercentage:
          type: number
        profitMarginPercentage:
          type: number
        discountPercentage:
          type: number
        vatEnabled:
          type: boolean
        vatRate:
          type: number
        toolingRecommendation:
          type: string
        materialRecommendation:
          type: string
        coolantRecommendation:
          type: string
    QuoteSummary:
      type: object
      required:
        - id
        - quoteNumber
        - customerId
        - customerName
        - status
        - quoteDate
        - validUntil
        - totalValue
        - createdAt
      properties:
        id:
          type: integer
        quoteNumber:
          type: string
        customerId:
          type: integer
        customerName:
          type: string
        status:
          type: string
        quoteDate:
          type: string
        validUntil:
          type: string
        totalValue:
          type: number
        createdAt:
          type: string
    Quote:
      type: object
      required:
        - id
        - quoteNumber
        - customerId
        - status
        - lostReason
        - quoteDate
        - validUntil
        - quoteRevision
        - revisionNotes
        - leadTime
        - deliveryTerms
        - notes
        - internalNotes
        - paymentTerms
        - termsAndConditions
        - materialCertIncluded
        - inspectionReportIncluded
        - fairIncluded
        - cmmReportIncluded
        - priceBreakQtys
        - lineItems
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        quoteNumber:
          type: string
        customerId:
          type: integer
        status:
          type: string
        lostReason:
          type: string
        quoteDate:
          type: string
        validUntil:
          type: string
        quoteRevision:
          type: string
        revisionNotes:
          type: string
        leadTime:
          type: string
        deliveryTerms:
          type: string
        notes:
          type: string
        internalNotes:
          type: string
        paymentTerms:
          type: string
        termsAndConditions:
          type: string
        materialCertIncluded:
          type: boolean
        inspectionReportIncluded:
          type: boolean
        fairIncluded:
          type: boolean
        cmmReportIncluded:
          type: boolean
        priceBreakQtys:
          type: string
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/QuoteLineItem"
        createdAt:
          type: string
        updatedAt:
          type: string
    QuoteInput:
      type: object
      required:
        - customerId
        - lineItems
      properties:
        customerId:
          type: integer
        status:
          type: string
        lostReason:
          type: string
        quoteDate:
          type: string
        validUntil:
          type: string
        quoteRevision:
          type: string
        revisionNotes:
          type: string
        leadTime:
          type: string
        deliveryTerms:
          type: string
        notes:
          type: string
        internalNotes:
          type: string
        paymentTerms:
          type: string
        termsAndConditions:
          type: string
        materialCertIncluded:
          type: boolean
        inspectionReportIncluded:
          type: boolean
        fairIncluded:
          type: boolean
        cmmReportIncluded:
          type: boolean
        priceBreakQtys:
          type: string
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/QuoteLineItemInput"
    QuoteUpdate:
      type: object
      properties:
        customerId:
          type: integer
        status:
          type: string
        lostReason:
          type: string
        quoteDate:
          type: string
        validUntil:
          type: string
        quoteRevision:
          type: string
        revisionNotes:
          type: string
        leadTime:
          type: string
        deliveryTerms:
          type: string
        notes:
          type: string
        internalNotes:
          type: string
        paymentTerms:
          type: string
        termsAndConditions:
          type: string
        materialCertIncluded:
          type: boolean
        inspectionReportIncluded:
          type: boolean
        fairIncluded:
          type: boolean
        cmmReportIncluded:
          type: boolean
        priceBreakQtys:
          type: string
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/QuoteLineItemInput"
...
    QuoteDrawing:
      type: object
      required:
        - id
        - quoteId
        - filename
        - contentType
        - objectPath
        - drawingNumber
        - revision
        - uploadedAt
      properties:
        id:
          type: integer
        quoteId:
          type: integer
        filename:
          type: string
        contentType:
          type: string
        objectPath:
          type: string
        drawingNumber:
          type: string
        revision:
          type: string
        uploadedAt:
          type: string
    CreateQuoteDrawingBody:
      type: object
      required:
        - objectPath
        - filename
        - contentType
      properties:
        objectPath:
          type: string
        filename:
          type: string
        contentType:
          type: string
        drawingNumber:
          type: string
        revision:
          type: string
...